requirements.txt
google-cloud-aiplatform
google-generativeai>=0.8.3
python-dotenv>=1.0.0
streamlit>=1.40.0
google-adk>=0.1.0 
duckduckgo-search
markdown2>=2.5.0
xhtml2pdf>=0.2.15